requirements.txt
streamlit
transformers
torch
torchvision
easyocr
python-docx
pillow
matplotlib